attached_assets/extracted/develop/requirements.txt
fastapi 
uvicorn 
httpx 
python-dotenv
pydantic
solcast
